ShowDoc: don't display header on empty docs
Fix issue #836 (`show-doc` "displays" nonexistent docs)

Instead of displaying a header without a comment, raise `CommandError`
and notify the user that a comment for the given definition doesn't
exist (but the very definition is present).

Add a test for this change. Also amend a couple of older tests, so they
conform the new behaviour.